Handle slow spawn.stdout
In a high demand case I've found times where spawn returned an
unfinished object, without in the returned object.

This patch provides a workaround for this problem, but at the
same time if should probably be fixed directly on spawn itself.
In the meanwhile, offers some compatibility over non up to date
Node versions.