d8: Remove experimental notice from stderr
If you invoke d8 with any --harmony flags, it prints a warning to stderr
that you are using experimental technologies. Remove this warning from
the captured stderr, because it's just noise - if you are running d8
with a --harmony flag via eshost, you probably know what you are doing.

Additionally, non-empty stderr makes test262harness think the test has
failed, so it's currently impossible to run test262 tests for a
--harmony feature.